make qna reactive
